requirements.txt
aniso8601==9.0.1
attrs==20.3.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
